requirements.txt
functions-framework==3.2.1
google-cloud-firestore==2.6.1
google-api-core==2.8.2
Flask==2.2.2
pika==1.3.2
firebase_admin==6.0.1
paho-mqtt==1.6.1